backend/requirements.txt
fastapi
firebase_admin
uvicorn
pymongo
celery
PyJWT